requirements.txt
psutil>=5.9
py-cpuinfo>=9.0
httpx>=0.24
jinja2>=3.1
Pillow>=10.0
